+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Fixed `/api/documentation/<package>` endpoint bug.
 
 ## `20220829t082800-all`
  * Bumped runtimeVersion to `2022.08.26`.